Rename map layer drag and drop to drag and projection, and move some of the pure projection logic out of the React hook and into a separate helper module. This better separates the concerns of (1) defining the stable IDs and commit logic for map layer dragging from (2) the React state management and hover behavior.
