+++ packages/itwin/map-layers/src/ui/widget/MapLayerDroppable.tsx
@@ -7,7 +7,7 @@
 import "./MapLayerDroppable.scss";
 import { MapLayersUI } from "../../mapLayers";
 import { AttachLayerButtonType, AttachLayerPopupButton } from "./AttachLayerPopupButton";
-import { backgroundMapLayersId, overlayMapLayersId } from "./MapLayerDragDrop";
+import { backgroundMapLayersId, overlayMapLayersId } from "./MapLayerDragModel";
 import { useDroppable } from "@dnd-kit/react";
 import { useMapLayerListContext } from "./MapLayerListContext";
